Add getVectorLengths to Foreign.Solver, and add getVectorStarts
Even though getVectorLengths is missing from CBC, it can be approximated using
getVectorStarts.  This is a bit of a lie though since they don't exactly
correspond.  So we'll do it the right way in Solver, and the wrong way in CBC.
The only bad thing is that calls to CBC's getElements may include some junk.